requirements.txt
openai
requests
python-dotenv
colorama
gTTS
asyncio
